web-app-frameworks/gradio-chat-app/requirements.txt
gradio==4.12.0
websocket-client==1.7.0